rename `sigma_proof::ctxt_ctxt_equality_proof` to non-abbreviated name
