+++ zk-token-sdk/src/sigma_proofs/ciphertext_ciphertext_equality_proof.rs
